+++ node/packages/aws-lambda-sdk/test/fixtures/lambdas/error-initialization.js
@@ -1,3 +1,3 @@
 'use strict';
 
-throw new Error('Initialization');
+throw new Error('Stop');