+++ node/packages/sdk/index.ts
@@ -1,4 +1,0 @@
-import { Sdk } from '@serverless/sdk';
-const sdk = require('./index.js');
-
-export default sdk as Sdk;