+++ node/packages/sdk/lib/error.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = class ServerlessSdkError extends Error {};