+++ node/packages/sdk/lib/trace-span.js
@@ -17,6 +17,8 @@ const generateId = require('./generate-id');
 const resolveEpochTimestampString = require('./resolve-epoch-timestamp-string');
 const toProtobufEpochTimestamp = require('./to-protobuf-epoch-timestamp');
 
+const objHasOwnProperty = Object.prototype.hasOwnProperty;
+
 class StringifiableSet extends Set {
   toJSON() {
     return Array.from(this);
@@ -25,11 +27,9 @@ class StringifiableSet extends Set {
 
 const asyncLocalStorage = new AsyncLocalStorage();
 
-let rootSpan;
-
 class TraceSpan {
   static resolveCurrentSpan() {
-    return asyncLocalStorage.getStore() || rootSpan || null;
+    return asyncLocalStorage.getStore() || TraceSpan.rootSpan || null;
   }
 
   constructor(name, options = {}) {
@@ -61,17 +61,19 @@ class TraceSpan {
     if (tags) this.tags.setMany(tags);
     if (options.input != null) this.input = options.input;
     if (options.output != null) this.output = options.output;
-    if (!rootSpan) {
-      rootSpan = this;
+    if (!TraceSpan.rootSpan) {
+      TraceSpan.rootSpan = this;
       this.parentSpan = null;
     } else {
-      if (rootSpan.endTime) {
+      if (TraceSpan.rootSpan.endTime) {
         throw Object.assign(new Error('Cannot intialize span: Trace is closed'), {
           code: 'UNREACHABLE_TRACE',
         });
       }
       this.parentSpan = TraceSpan.resolveCurrentSpan();
-      while (this.parentSpan.endTime) this.parentSpan = this.parentSpan.parentSpan || rootSpan;
+      while (this.parentSpan.endTime) {
+        this.parentSpan = this.parentSpan.parentSpan || TraceSpan.rootSpan;
+      }
     }
 
     asyncLocalStorage.enterWith(this);
@@ -88,13 +90,13 @@ class TraceSpan {
   }
   closeContext() {
     if (asyncLocalStorage.getStore() !== this) return;
-    if (this === rootSpan) {
+    if (this === TraceSpan.rootSpan) {
       asyncLocalStorage.enterWith(this);
       return;
     }
     const openParentSpan = (function self(span) {
       if (!span.endTime) return span;
-      return span.parentSpan ? self(span.parentSpan) : rootSpan;
+      return span.parentSpan ? self(span.parentSpan) : TraceSpan.rootSpan;
     })(this.parentSpan);
     asyncLocalStorage.enterWith(openParentSpan);
   }
@@ -121,7 +123,7 @@ class TraceSpan {
       }
     }
     this.endTime = targetEndTime || defaultEndTime;
-    if (this === rootSpan) {
+    if (this === TraceSpan.rootSpan) {
       const leftoverSpans = [];
       for (const subSpan of this.spans) {
         if (subSpan.endTime) continue;
@@ -157,6 +159,7 @@ class TraceSpan {
       input: this.input || undefined,
       output: this.output || undefined,
       tags: this.tags,
+      customTags: objHasOwnProperty.call(this, 'customTags') ? this.customTags : undefined,
     };
   }
   toProtobufObject() {
@@ -170,6 +173,9 @@ class TraceSpan {
       input: this.input || undefined,
       output: this.output || undefined,
       tags: toProtobufTags(this.tags),
+      customTags: objHasOwnProperty.call(this, 'customTags')
+        ? JSON.stringify(this.customTags)
+        : undefined,
     };
   }
   get spans() {
@@ -205,6 +211,7 @@ Object.defineProperties(
     id: d(() => generateId()),
     subSpans: d(() => new Set()),
     tags: d(() => new Tags()),
+    customTags: d(() => new Tags()),
   })
 );