+++ node/packages/sdk/test/unit/lib/instrumentation/node-console.test.js
@@ -34,4 +34,16 @@ describe('lib/instrumentation/node-console.js', () => {
     expect(capturedEvent.tags.get('warning.message')).to.equal('My message 12 true');
     expect(capturedEvent._origin).to.equal('nodeConsole');
   });
+
+  it('should recognize Serverless SDK warning', () => {
+    let capturedEvent;
+    serverlessSdk._eventEmitter.once('captured-event', (event) => (capturedEvent = event));
+    // eslint-disable-next-line no-console
+    console.warn({ source: 'serverlessSdk', message: 'Something is wrong' });
+
+    expect(capturedEvent.name).to.equal('telemetry.warning.generated.v1');
+    expect(capturedEvent.tags.get('warning.message')).to.equal('Something is wrong');
+    expect(capturedEvent.tags.get('warning.type')).to.equal(2);
+    expect(capturedEvent._origin).to.equal('nodeConsole');
+  });
 });