+++ node/packages/aws-lambda-sdk/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.13.1](https://github.com/serverless/console/compare/@serverless/aws-lambda-sdk@0.13.0...@serverless/aws-lambda-sdk@0.13.1) (2023-01-13)
+
+### Features
+
+- Report traces at unhandled exceptions in invocation phase ([eadeaee](https://github.com/serverless/console/commit/eadeaeedb04390c101a7b2f0d9eae2acd317e0d1))
+- Upgrade to `@serverless/sdk` v0.3.1 (Introduce `servelressSdk.setTag`) ([78c5532](https://github.com/serverless/console/commit/78c5532b13588f27d84b3a2acdfc9b0e1987fd4f))
+
+### Maintenance Improvements
+
+- Support trace closure at initialization phase ([48262d0](https://github.com/serverless/console/commit/48262d0962889fe9fb5eae75b8c329fb00551f5e))
+
+## [0.13.0](https://github.com/serverless/console/compare/@serverless/aws-lambda-sdk@0.12.0...@serverless/aws-lambda-sdk@0.13.0) (2023-01-12)
+
+### ⚠ BREAKING CHANGES
+
+- Upgrade to `@serverless/sdk` v0.3. `serverlessSdk.captureTraceSpan` is no longer available
+
+### Features
+
+- Upgrade to `@serverless/sdk` v0.3 ([e415fbc](https://github.com/serverless/console/commit/e415fbcdb8b80cd0c0fb6bffe754d4f654bf66ca))
+- Report internal warnings with structured logs ([63d046a](https://github.com/serverless/console/commit/63d046a033565d1196d2d1ae89fbae81e938e226))
+- Ensure DynamoDB tags on AWS SDK spans that cover DynamoDbDocument service ([#354](https://github.com/serverless/console/issues/354)) ([1b49f9e](https://github.com/serverless/console/commit/1b49f9e9affecfcf469dc469bb43470bdedcfc41))
+
 ## [0.12.0](https://github.com/serverless/console/compare/@serverless/aws-lambda-sdk@0.11.6...@serverless/aws-lambda-sdk@0.12.0) (2023-01-05)
 
 ### ⚠ BREAKING CHANGES